pyproject.toml
[tool.poetry]
name = "highstreets"
version = "0.1.0"
description = "Analysis and modeling of London high street profiles."
authors = ["Conor Dempsey <conor.dempsey@london.gov.uk>"]
license = "MIT"
repository = "https://github.com/Greater-London-Authority/highstreets"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8,<3.11"
pandas = "^1.4.2"
dill = "^0.3.4"
ipykernel = "^6.13.0"
ipython = "^8.2.0"
jupyter = "^1.0.0"
matplotlib = "^3.5.1"
numpy = "^1.22.3"
python-dotenv = "^0.20.0"
scikit-learn = "^1.0.2"
scipy = "^1.8.0"
seaborn = "^0.11.2"
openpyxl = "^3.0.9"
xgboost = "^1.6.1"
sqlalchemy = "^2.0.4"
psycopg2-binary = "^2.9.5"
pydantic = "1.9"
pandera = "^0.13.4"
tqdm = "^4.64.1"
requests = "^2.27.1"
geopandas = "^0.9.0"
geoalchemy2 = "^0.13.3"
psycopg2 = "^2.8.6"
glapy = {git = "https://github.com/Greater-London-Authority/glapy"}
shapely = "1.8.5"
pygeos = "^0.14"

[tool.poetry.dev-dependencies]
debugpy = "^1.6.0"
pre-commit = "^2.18.1"

[build-system]
requires = ["poetry-core>=1.0.0", "setuptools>=57.0.0"]
build-backend = "poetry.core.masonry.api"
